feat: creativity loss

+++ src/dataloaders/load_data.py
@@ -10,11 +10,11 @@ def load_data(config: Config, img_target_shape: Tuple[int, int], embed_data: boo
     if config.name == 'CUB':
         ds_train = cub.load_dataset(config.dir, is_train=True, target_shape=img_target_shape)
         ds_test = cub.load_dataset(config.dir, is_train=False, target_shape=img_target_shape)
-        class_attributes = cub.load_class_attributes(config.dir)
+        class_attributes = cub.load_class_attributes(config.dir).astype(np.float32)
     elif config.name == 'AWA':
         ds_train = awa.load_dataset(config.dir, split='train', target_shape=img_target_shape)
         ds_test = awa.load_dataset(config.dir, split='test', target_shape=img_target_shape)
-        class_attributes = awa.load_class_attributes(config.dir)
+        class_attributes = awa.load_class_attributes(config.dir).astype(np.float32)
     elif config.name == 'SVHN':
         ds_train = svhn.load_dataset(config.dir, split='train', target_shape=img_target_shape)
         ds_test = svhn.load_dataset(config.dir, split='test', target_shape=img_target_shape)
